refactor traverse() logic into extract-ast-nodes
